Refactor the infrastructure directories
