--- src/main/java/com/orojas/finance/infrastructure/adapters/output/persistence/repository/PaymentRepository.java
@@ -1,9 +1,0 @@
-package com.orojas.finance.infrastructure.adapters.output.persistence.repository;
-
-import com.orojas.finance.infrastructure.adapters.output.persistence.entity.PaymentEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface PaymentRepository extends JpaRepository<PaymentEntity, UUID> {
-}
