+++ src/main/java/com/orojas/finance/infrastructure/persistence/repository/PaymentRepository.java
@@ -1,0 +1,9 @@
+package com.orojas.finance.infrastructure.persistence.repository;
+
+import com.orojas.finance.infrastructure.persistence.entity.PaymentEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface PaymentRepository extends JpaRepository<PaymentEntity, UUID> {
+}
