+++ src/main/java/com/orojas/finance/infrastructure/rest/mapper/PaymentRestMapper.java
@@ -1,8 +1,8 @@
-package com.orojas.finance.infrastructure.adapters.input.rest.mapper;
+package com.orojas.finance.infrastructure.rest.mapper;
 
 import com.orojas.finance.domain.model.Payment;
-import com.orojas.finance.infrastructure.adapters.input.rest.model.request.PaymentRequest;
-import com.orojas.finance.infrastructure.adapters.input.rest.model.response.PaymentResponse;
+import com.orojas.finance.infrastructure.rest.model.request.PaymentRequest;
+import com.orojas.finance.infrastructure.rest.model.response.PaymentResponse;
 import org.mapstruct.Mapper;
 
 import java.util.List;
